Resolve corrupted flash problems with ARDS...
* ARDS seems to get corrupted data after flashing past 512KB. For now
flash size is changed to match GnM until the problem causing this is
found.
* Write function improved. Instead of loading entire flash file into ram
(which could become an issue on DS/DS Lite consoles if/when ARDS is able
to use 2MB flash dumps again), the file is read into smaller buffer and
read/written to in the same size chunks that the main write loop
operates in.