Hide unsupported card text on boot
* If used as an installed app on DSi System Menu (via HiyaCFW) the
bottom screen may display "unsupported cart" message on bottom screen
during initial bootup and card init despite having a supproted cart
inserted. Once card init finishes this message goes away. To prevent
confusion I have added a one time flag that is set to false on boot
after card init runs at least once so that this message is not displayed
until card init runs at least once and the app knows if it's supported
or not. This does did not impact Unlaunch as the card was already inited
properly before the app boots where as System Menu doesn't see most
early Datel carts as valid and powers the slot down when booting DSiWare
which was the reason behind the console initially showing this text
while it was doing initial slot power on and init.